mcp_client/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
beartype==0.22.5
blinker==1.9.0
cachetools==6.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.3
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.13.0.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
langchain==1.0.5
langchain-core==1.0.4
langchain-mcp-adapters==0.1.13
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.21.0
mdurl==0.1.2
more-itertools==10.8.0
msal==1.34.0
narwhals==2.11.0
numpy==2.3.4
openai==2.7.2
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.0
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
streamlit==1.51.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
